requirements.txt
raspberrypi-python-tm1637==1.3.0
RPi.GPIO==0.7.0
wiringpi==2.60.0
